+++ itest/babylon/babylon_test_manager.go
@@ -200,7 +200,6 @@ func (tm *TestManager) AddFinalityProvider(t *testing.T, ctx context.Context, hm
 	cfg.BabylonConfig.Key = fpKeyName
 	cfg.BabylonConfig.RPCAddr = fmt.Sprintf("http://localhost:%s", tm.babylond.GetPort("26657/tcp"))
 	cfg.BabylonConfig.GRPCAddr = fmt.Sprintf("https://localhost:%s", tm.babylond.GetPort("9090/tcp"))
-	cfg.ContextSigningHeight = ^uint64(0) // max uint64 to enable context signing
 
 	// Set HMAC key if provided
 	if len(hmacKey) > 0 && hmacKey[0] != "" {
@@ -413,6 +412,9 @@ func (tm *TestManager) StopAndRestartFpAfterNBlocks(ctx context.Context, t *test
 
 	err = fpIns.Start(ctx)
 	require.NoError(t, err)
+
+	// Add sleep to allow database to initialize
+	time.Sleep(15 * time.Second)
 }
 
 func (tm *TestManager) WaitForNBlocks(t *testing.T, n int) uint64 {